DeepOpsBackend/requirements.txt
django
django-admin
django-rest-framework
django_filter
drf-spectacular
drf-spectacular-sidecar
environ